Added units to GUI map
This is temporary and needs to be restructured, since I have discovered it is not possible (or not easy) to draw components on top of each other. This is why the tiles and units are side-by-side instead of on top of each other. I will create a new component that draws them on top of each other, but it will require changing how mouse clicks are handled. For now, this allows tile- and unit-selection though.