Moving around the map is functional!
Extracted a lot of MapPanel logic to a new model, MapOverlay. This provides a model for the player's currently selected units and tiles on the map. I still need to make MapPanel, MapView, GameView, and MapController conform to MVC standards better, since MapController is sort of acting as a GameController right now.

With the logic moved into MapOverlay and MapController, movements and attacks have been added, as well as ending a turn (to allow another movement to be performed, etc.). Attacks and movements are functional for determining valid locations, but attacks are not yet working for actually dealing damage.

Fixed a bug where tiles were being added/accessed with X and Y swapped. Keep in mind that Locations are (x, y), but 2D arrays in programming are [row][column], i.e. [y][x].

Fixed a bug where moving a unit did not actually update it's location on the map. To fix this, Map's now filter a stream of units by location when searching for them rather than using a hashmap (which is not optimal, but doesn't matter for how few units we have at any time).